Create hook to fetch next page results
