+++ pages/index.tsx
@@ -3,27 +3,13 @@ import { useFetchPokemonWithInfinityScroll } from './hooks/usePokemon';
 import HomePage from '@/components/pages/HomePage';
 import ContainerW1024 from '@/components/UI/atoms/ContainerW1024';
 import MyBeatLoader from '@/components/UI/molecules/BeatLoader';
+import useFetchNextPage from './hooks/useFetchNextPage';
 
 const IndexPage: FC = (): JSX.Element => {
   const { data, isSuccess, hasNextPage, fetchNextPage, isInitialLoading } =
     useFetchPokemonWithInfinityScroll();
 
-  useEffect(() => {
-    let fetching = false;
-    const handleScroll = async (e: any) => {
-      const { scrollHeight, scrollTop, clientHeight } =
-        e.target.scrollingElement;
-      if (!fetching && scrollHeight - scrollTop <= clientHeight * 1.2) {
-        fetching = true;
-        if (hasNextPage) await fetchNextPage();
-        fetching = false;
-      }
-    };
-    document.addEventListener('scroll', handleScroll);
-    return () => {
-      document.removeEventListener('scroll', handleScroll);
-    };
-  }, [fetchNextPage, hasNextPage]);
+  useFetchNextPage(hasNextPage, fetchNextPage);
 
   if (isInitialLoading) {
     return (
